fix go.mod's
