+++ extension/basicauth/basicauth.go
@@ -16,8 +16,10 @@ package basicauth // import "github.com/open-telemetry/opentelemetry-collector-c
 
 import (
 	"context"
+	"encoding/base64"
 	"errors"
 	"fmt"
+	"strings"
 
 	"github.com/tg123/go-htpasswd"
 	"go.opentelemetry.io/collector/client"
@@ -48,8 +50,10 @@ func (ba *BasicAuth) Shutdown(ctx context.Context) error {
 }
 
 var (
-	errNoAuth             = errors.New("no basic auth provided")
-	errInvalidCredentials = errors.New("invalid credentials")
+	errNoAuth              = errors.New("no basic auth provided")
+	errInvalidCredentials  = errors.New("invalid credentials")
+	errInvalidSchemePrefix = errors.New("invalid authorization scheme prefix")
+	errInvalidFormat       = errors.New("invalid authorization format")
 )
 
 func (ba *BasicAuth) Authenticate(ctx context.Context, headers map[string][]string) (context.Context, error) {
@@ -70,4 +74,53 @@ func (ba *BasicAuth) Authenticate(ctx context.Context, headers map[string][]stri
 	cl := client.FromContext(ctx)
 	cl.Auth = authData
 	return client.NewContext(ctx, cl), nil
+}
+
+// See: https://github.com/golang/go/blob/1a8b4e05b1ff7a52c6d40fad73bcad612168d094/src/net/http/request.go#L950
+func parseBasicAuth(auth string) (*authData, error) {
+	const prefix = "Basic "
+	if len(auth) < len(prefix) || !strings.EqualFold(auth[:len(prefix)], prefix) {
+		return nil, errInvalidSchemePrefix
+	}
+
+	encoded := auth[len(prefix):]
+	decodedBytes, err := base64.StdEncoding.DecodeString(encoded)
+	if err != nil {
+		return nil, errInvalidFormat
+	}
+	decoded := string(decodedBytes)
+
+	si := strings.IndexByte(decoded, ':')
+	if si < 0 {
+		return nil, errInvalidFormat
+	}
+
+	return &authData{
+		username: decoded[:si],
+		password: decoded[si+1:],
+		raw:      encoded,
+	}, nil
+}
+
+var _ client.AuthData = (*authData)(nil)
+
+type authData struct {
+	username string
+	password string
+	raw      string
+}
+
+func (a *authData) GetAttribute(name string) interface{} {
+	switch name {
+	case "subject":
+		return a.username
+	case "raw":
+		return a.raw
+	default:
+		return nil
+	}
+}
+
+func (*authData) GetAttributeNames() []string {
+	return []string{"subject", "raw"}
 }
